+++ src/views/HomeView.vue
@@ -28,7 +28,8 @@
   imports
 */
 
-  import { ref, reactive, computed, watch, onMounted, nextTick } from 'vue'
+  import { ref, onMounted } from 'vue'
+  import { useCounter } from '@/use/useCounter'
   import { vAutofocus } from '@/directives/vAutofocus'
 
 /*
@@ -47,35 +48,9 @@
   counter
 */
 
-  const counterData = reactive({
-    count: 0,
-    title: 'My Counter'
-  })
+  const { counterData, oddOrEven, increaseCounter, decreaseCounter } = useCounter()
 
-  watch(() => counterData.count, (newCount) => {
-    if (newCount === 20) {
-      alert('Way to go! You made it to 20!!')
-    }
-  })
 
-  const oddOrEven = computed(() => {
-    if (counterData.count % 2 === 0) return 'even'
-    return 'odd'
-  })
-
-  const increaseCounter = async (amount, e) => {
-    counterData.count += amount
-    await nextTick()
-    console.log('do something when counter has updated in the dom')
-  }
-
-  const decreaseCounter = amount => {
-    counterData.count -= amount
-  }
-
-  onMounted(() => {
-    console.log('Do stuff related to Counter')
-  })
 
 </script>
