+++ src/views/PostsView.vue
@@ -12,6 +12,16 @@
     </ul>
 
     <textarea v-autofocus />
+
+    <div>
+      <button
+        @click="increaseCounter(1)"
+        class="counter-button"
+        :class="{ 'yellow' : oddOrEven === 'odd' }"
+      >
+        {{ counterData.count }}
+      </button>
+    </div>
   </div>
 </template>
 
@@ -21,6 +31,7 @@
 */
 
   import { ref } from 'vue'
+  import { useCounter } from '@/use/useCounter'
   import { vAutofocus } from '@/directives/vAutofocus'
 
 /*
@@ -42,10 +53,24 @@
     }
   ])
 
+/*
+  counter button
+*/
+
+  const { counterData, increaseCounter, oddOrEven } = useCounter()
+
 </script>
 
 <style scoped>
 ul {
   margin-bottom: 30px;
 }
+.counter-button {
+  font-size: 60px;
+  width: 100%;
+  background-color: pink;
+}
+.counter-button.yellow {
+  background-color: yellow;
+} 
 </style>
